myproject/requirements.txt
arrow==1.2.3
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
colorama==0.4.6
cookiecutter==2.1.1
crispy-bootstrap4==2022.1
dj-database-url==1.2.0
Django==4.1.3
django-browser-reload==1.6.0
django-crispy-forms==2.0
django-extensions==3.0.2
django-tailwind==3.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
graphviz==0.14
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.20
mysqlclient==2.1.1
packaging==23.1
Pillow==9.3.0
protobuf==3.12.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pydot==1.4.1
pygame==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.6
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stripe==6.5.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
